Move up conversion into CertRevocationListError
